refactor: seperate login for teacher and student
