+++ frontend/src/app/pages/student/StudentLogin.tsx
@@ -1,0 +1,8 @@
+import AuthPage from "@/components/Auth/AuthPage"
+
+type Role = 'teacher' | 'student'
+
+const StudentLogin = () => <AuthPage role={'student' as Role} />
+
+
+export default StudentLogin
