+++ frontend/src/app/pages/teacher/TeacherLogin.tsx
@@ -1,0 +1,7 @@
+import AuthPage from "@/components/Auth/AuthPage"
+
+type Role = 'teacher' | 'student'
+
+const TeacherLogin = () => <AuthPage role={'teacher' as Role} />
+
+export default TeacherLogin
